// src/client/service/enums/access.ts
export enum AccessEnum {
    DICTIONARY = "DICTIONARY",
    DICTIONARY_USERS = "DICTIONARY_USERS",
    DICTIONARY_BRAND = "DICTIONARY_BRAND",
    DICTIONARY_MODELS = "DICTIONARY_MODELS",
    DICTIONARY_SERVICE = "DICTIONARY_SERVICE",
    DICTIONARY_CARPART = "DICTIONARY_CARPART",
    PROPOSAL = "PROPOSAL",
    PROPOSAL_ALL = "PROPOSAL_ALL",
    PROPOSAL_ADD = "PROPOSAL_ADD",
    PROPOSAL_CLIENTS = "PROPOSAL_CLIENTS",
    PROPOSAL_CARS = "PROPOSAL_CARS",
    STATISTIC = "STATISTIC",
    STATISTIC_EVERYDAY = "STATISTIC_EVERYDAY",
}